Refactor return value for the clock collecting
Previously it was a dictionary with the key being the "kind" of
clocks, e.g. "deadline", "scheduled", etc. mapping to a list of tuples
containing the clock information, now it is a single list with
dictionary elements instead of tuples, which has a kind attribute to
filter it.
This is much easier to work with, as the dictionary can be extended to
contain more information without many pieces of code having to know
about the tuple structure.